# docs/swagger.yaml
basePath: /v1
definitions:
  main.Task:
    properties:
      command:
        type: string
      created_at:
        type: string
      error:
        type: string
      id:
        type: string
      output:
        type: string
      status:
        $ref: '#/definitions/main.TaskStatus'
      updated_at:
        type: string
    type: object
  main.TaskStatus:
    enum:
    - pending
    - running
    - finished
    - failed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusRunning
    - StatusFinished
    - StatusFailed
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing the kernel service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kernel Service API
  version: "1.0"
paths:
  /restart:
    post:
      description: Restarts the kernel service asynchronously
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Restart kernel service
      tags:
      - Kernel
  /set-domain:
    post:
      consumes:
      - application/json
      description: Sets the domain for the kernel service asynchronously
      parameters:
      - description: Domain to set
        in: body
        name: domain
        required: true
        schema:
          properties:
            domain:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set domain for kernel service
      tags:
      - Configuration
  /set-ip:
    post:
      consumes:
      - application/json
      description: Sets the IP plan for the kernel service asynchronously
      parameters:
      - description: IPv4 address to set
        in: body
        name: ipv4
        required: true
        schema:
          properties:
            ipv4:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set IP plan for kernel service
      tags:
      - Configuration
  /start:
    post:
      description: Starts the kernel service asynchronously
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start kernel service
      tags:
      - Kernel
  /status:
    get:
      description: Returns the current status of the kernel service (asynchronous)
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get kernel service status
      tags:
      - Kernel
  /stop:
    post:
      description: Stops the kernel service asynchronously
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stop kernel service
      tags:
      - Kernel
  /task/{id}:
    get:
      description: Returns the current status of a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
      summary: Get task status
      tags:
      - Task
swagger: "2.0"